requirements.txt
streamlit-audiorec
streamlit-webrtc
PyPDF2
langchain
langchain_google_genai
google-generativeai
langchain_community
faiss-cpu
requests
python-dotenv
python-docx
pytesseract
Pillow
pdf2image
selenium
speechrecognition
gTTS
bs4
